fix: add missing exports for vaadin-messages
